# MessengerWithWheaterApp/dockerGuziec/guziec/apiclient/model/message.ts
/**
 * ZiemniaczaneRozmowy
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Message { 
    content: string;
    author_id: number;
    id: number;
    reciver_id: number;
    is_recived: boolean;
    date: string;
}