# MessengerWithWheaterApp/dockerGuziec/guziec/apiclient/model/messageCreate.ts
/**
 * ZiemniaczaneRozmowy
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MessageCreate { 
    content: string;
    author_id: number;
}