# MessengerWithWheaterApp/dockerGuziec/guziec/openapi.json
{"openapi":"3.0.2","info":{"title":"ZiemniaczaneRozmowy","version":"1.0.0"},"paths":{"/user/":{"get":{"summary":"Read Users","operationId":"read_users_user__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Users User  Get","type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create User","operationId":"create_user_user__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/login/":{"post":{"summary":"Login User","operationId":"login_user_user_login__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/logout/":{"post":{"summary":"Logout User","operationId":"logout_user_user_logout__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/status":{"put":{"summary":"Update User Status","operationId":"update_user_status_users_status_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/messages/{reciver}/{author}/":{"get":{"summary":"Get Messages To User","operationId":"get_messages_to_user_user_messages__reciver___author___get","parameters":[{"required":true,"schema":{"title":"Reciver","type":"integer"},"name":"reciver","in":"path"},{"required":true,"schema":{"title":"Author","type":"integer"},"name":"author","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Messages To User User Messages  Reciver   Author   Get","type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{receiver_id}/{sender_id}/":{"get":{"summary":"Read Messages To User From","operationId":"read_messages_to_user_from_message__receiver_id___sender_id___get","parameters":[{"required":true,"schema":{"title":"Receiver Id","type":"integer"},"name":"receiver_id","in":"path"},{"required":true,"schema":{"title":"Sender Id","type":"integer"},"name":"sender_id","in":"path"},{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Messages To User From Message  Receiver Id   Sender Id   Get","type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/general/":{"get":{"summary":"Read Messages General","operationId":"read_messages_general_general__get","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Messages General General  Get","type":"array","items":{"$ref":"#/components/schemas/Message"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/singleton//{receiver_id}/{sender_id}/":{"get":{"summary":"Get Singleton","operationId":"get_singleton_singleton___receiver_id___sender_id___get","parameters":[{"required":true,"schema":{"title":"Receiver Id","type":"integer"},"name":"receiver_id","in":"path"},{"required":true,"schema":{"title":"Sender Id","type":"integer"},"name":"sender_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/messages/":{"post":{"summary":"Create Message","operationId":"create_message_user_messages__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_create_message_user_messages__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{reciver}/read":{"put":{"summary":"Update Message Status","operationId":"update_message_status_message__reciver__read_put","parameters":[{"required":true,"schema":{"title":"Reciver","type":"integer"},"name":"reciver","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Messages Ids","type":"array","items":{}}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/message/{reciver}/{author}/unread":{"post":{"summary":"Unread Messages To User From","operationId":"unread_messages_to_user_from_message__reciver___author__unread_post","parameters":[{"required":true,"schema":{"title":"Reciver","type":"integer"},"name":"reciver","in":"path"},{"required":true,"schema":{"title":"Author","type":"integer"},"name":"author","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_create_message_user_messages__post":{"title":"Body_create_message_user_messages__post","required":["message","reciver"],"type":"object","properties":{"message":{"$ref":"#/components/schemas/MessageCreate"},"reciver":{"$ref":"#/components/schemas/UserBase"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Message":{"title":"Message","required":["content","author_id","id","reciver_id","is_recived","date"],"type":"object","properties":{"content":{"title":"Content","type":"string"},"author_id":{"title":"Author Id","type":"integer"},"id":{"title":"Id","type":"integer"},"reciver_id":{"title":"Reciver Id","type":"integer"},"is_recived":{"title":"Is Recived","type":"boolean"},"date":{"title":"Date","type":"string","format":"date-time"}}},"MessageCreate":{"title":"MessageCreate","required":["content","author_id"],"type":"object","properties":{"content":{"title":"Content","type":"string"},"author_id":{"title":"Author Id","type":"integer"}}},"User":{"title":"User","required":["login","id","is_active"],"type":"object","properties":{"login":{"title":"Login","type":"string"},"id":{"title":"Id","type":"integer"},"is_active":{"title":"Is Active","type":"boolean"}}},"UserBase":{"title":"UserBase","required":["login"],"type":"object","properties":{"login":{"title":"Login","type":"string"}}},"UserCreate":{"title":"UserCreate","required":["login","password"],"type":"object","properties":{"login":{"title":"Login","type":"string"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}